refactor: 자주쓰는 Mock 함수들 util화
