+++ src/testUtil/TestUtil.js
@@ -1,0 +1,30 @@
+import { MissionUtils } from "@woowacourse/mission-utils";
+
+export const mockQuestions = (inputs) => {
+  const messages = [];
+
+  MissionUtils.Console.readLineAsync = jest.fn((prompt) => {
+    messages.push(prompt);
+    const input = inputs.shift();
+
+    if (input === undefined) {
+      throw new Error("NO INPUT");
+    }
+
+    return Promise.resolve(input);
+  });
+
+  MissionUtils.Console.readLineAsync.messages = messages;
+};
+
+export const mockNowDate = (date = null) => {
+  const mockDateTimes = jest.spyOn(MissionUtils.DateTimes, "now");
+  mockDateTimes.mockReturnValue(new Date(date));
+  return mockDateTimes;
+};
+
+export const getLogSpy = () => {
+  const logSpy = jest.spyOn(MissionUtils.Console, "print");
+  logSpy.mockClear();
+  return logSpy;
+};
